+++ Runtime/NestedEventHandlerTable.cs
@@ -3,16 +3,16 @@ using System.Collections.Generic;
 
 namespace DJM.EventManager
 {
-    public class NestedEventTable<TOuterKey, TInnerKey>
+    public class NestedEventHandlerTable<TOuterKey, TInnerKey>
     {
         private readonly Dictionary<TOuterKey, Dictionary<TInnerKey, Delegate>> _outerDict;
 
-        public NestedEventTable()
+        public NestedEventHandlerTable()
         {
             _outerDict = new Dictionary<TOuterKey, Dictionary<TInnerKey, Delegate>>();
         }
 
-        public Delegate GetValue(TOuterKey outerKey, TInnerKey innerKey)
+        public Delegate GetHandlers(TOuterKey outerKey, TInnerKey innerKey)
         {
             if (!_outerDict.TryGetValue(outerKey, out var innerDict))
                 return default(Delegate);
@@ -20,48 +20,65 @@ namespace DJM.EventManager
             return innerDict.TryGetValue(innerKey, out var dictValue) ? dictValue : default(Delegate);
         }
         
-        public void Add(TOuterKey outerKey, TInnerKey innerKey, Delegate value)
+        public void AddHandler(TOuterKey outerKey, TInnerKey innerKey, Delegate handler)
         {
             if (!_outerDict.TryGetValue(outerKey, out var innerDict))
             {
-                _outerDict[outerKey] = new Dictionary<TInnerKey, Delegate> {[innerKey] = value};
+                _outerDict[outerKey] = new Dictionary<TInnerKey, Delegate> {[innerKey] = handler};
                 return;
             }
 
             if (!innerDict.TryGetValue(innerKey, out var innerValue))
             {
-                innerDict[innerKey] = value;
+                innerDict[innerKey] = handler;
                 return;
             }
             
-            innerDict[innerKey] = Delegate.Combine(innerValue, value);
+            innerDict[innerKey] = Delegate.Combine(innerValue, handler);
         }
         
-        public void Remove(TOuterKey outerKey, TInnerKey innerKey, Delegate value)
+        public void RemoveHandler(TOuterKey outerKey, TInnerKey innerKey, Delegate handler)
         {
             if (!_outerDict.TryGetValue(outerKey, out var innerDict))
                 return;
             
-            if (!innerDict.TryGetValue(innerKey, out var innerValue))
+            if (!innerDict.TryGetValue(innerKey, out var handlers))
                 return;
             
-            innerDict[innerKey] = Delegate.Remove(innerValue, value);
+            var result = Delegate.Remove(handlers, handler);
+
+            if (result is null)
+            {
+                innerDict.Remove(innerKey);
+
+                if (innerDict.Count == 0)
+                    ClearEvent(outerKey);
+            }
+            else
+            {
+                innerDict[innerKey] = result;
+            }
         }
 
-        public void Clear(TOuterKey outerKey, TInnerKey innerKey)
+        public void ClearHandlers(TOuterKey outerKey, TInnerKey innerKey)
         {
             if (!_outerDict.TryGetValue(outerKey, out var innerDict))
                 return;
 
             innerDict.Remove(innerKey);
+
+            if (innerDict.Count == 0)
+            {
+                ClearEvent(outerKey);
+            }
         }
         
-        public void Clear(TOuterKey outerKey)
+        public void ClearEvent(TOuterKey outerKey)
         {
             _outerDict.Remove(outerKey);
         }
 
-        public void Clear()
+        public void ClearAll()
         {
             _outerDict.Clear();
         }